moved vcount, lcd_stat into top-level video obj
