+++ spa/src/common/videomem/mod.rs
@@ -9,7 +9,6 @@ pub use palette::PaletteRAM;
 pub use registers::*;
 pub use vram::VRAM2D;
 
-// TODO: make generic for VRAM
 pub struct VideoMemory<V: VRAM2D> {
     pub registers:  VideoRegisters,
